# internal/server/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Mockup server API",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b> module."
  },
  "host": "localhost:2327",
  "basePath": "/",
  "schemes": [
    "http",
    "https"
  ],
  "paths": {
    "/doc/swagger.json": {
      "get": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/": {
      "get": {
        "description": "",
        "parameters": [],
        "responses": {}
      }
    },
    "/detect": {
      "post": {
        "description": "",
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "name": "image_file",
            "in": "formData",
            "type": "file",
            "required": true,
            "description": "Upload image"
          },
          {
            "name": "split4",
            "in": "body",
            "description": "Chia thành 4 hình và detect , Mặc định set bằng true",
            "schema": {
              "type": "object",
              "properties": {
                "split4": {
                  "type": "boolean",
                  "example": true
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successfully obtained",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "number",
                  "example": 0
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "boxes": {
                      "type": "array",
                      "items": {
                        "type": "array",
                        "example": [
                          1051.6388397216797,
                          660.6006574630737,
                          1096.4102325439453,
                          721.9072008132935,
                          "8C",
                          0.6215610504150391
                        ],
                        "items": {
                          "type": "number"
                        }
                      }
                    },
                    "totalObjects": {
                      "type": "number",
                      "example": 1
                    }
                  }
                },
                "runTime": {
                  "type": "number",
                  "example": 0.285
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/detect-split4": {
      "post": {
        "description": "",
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "name": "image_file",
            "in": "formData",
            "type": "file",
            "required": true,
            "description": "Upload image"
          }
        ],
        "responses": {
          "200": {
            "description": "successfully obtained",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "number",
                  "example": 0
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "boxes": {
                      "type": "array",
                      "items": {
                        "type": "array",
                        "example": [
                          1051.6388397216797,
                          660.6006574630737,
                          1096.4102325439453,
                          721.9072008132935,
                          "8C",
                          0.6215610504150391
                        ],
                        "items": {
                          "type": "number"
                        }
                      }
                    },
                    "totalObjects": {
                      "type": "number",
                      "example": 1
                    }
                  }
                },
                "runTime": {
                  "type": "number",
                  "example": 0.285
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/sort-chinese-poker": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "cards",
            "in": "body",
            "description": "13 cards names",
            "schema": {
              "type": "array",
              "example": [
                "Ac",
                "2s",
                "2h",
                "2d",
                "2c",
                "3h",
                "4h",
                "5c",
                "7h",
                "7d",
                "9h",
                "10d",
                "Jh"
              ],
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": " successfully obtained",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "number",
                  "example": 0
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "totalRoyalty": {
                      "type": "number",
                      "example": 18
                    },
                    "chinesePokerSort": {
                      "type": "object",
                      "properties": {
                        "top": {
                          "type": "array",
                          "example": [
                            "AC",
                            "7D",
                            "5C"
                          ],
                          "items": {
                            "type": "string"
                          }
                        },
                        "mid": {
                          "type": "array",
                          "example": [
                            "JH",
                            "9H",
                            "7H",
                            "4H",
                            "3H"
                          ],
                          "items": {
                            "type": "string"
                          }
                        },
                        "bottom": {
                          "type": "array",
                          "example": [
                            "10D",
                            "2S",
                            "2H",
                            "2D",
                            "2C"
                          ],
                          "items": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                },
                "runTime": {
                  "type": "number",
                  "example": 180
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    }
  }
}